% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_features}
\alias{add_features}
\title{Add Features to Feature Layer}
\usage{
add_features(
  x,
  .data,
  rollback_on_fail = TRUE,
  token = Sys.getenv("ARCGIS_TOKEN")
)
}
\arguments{
\item{x}{an object of class \code{FeatureLayer}}

\item{.data}{an object of class \code{sf} or \code{data.frame}}

\item{rollback_on_fail}{if anything errors, roll back writes. Defaults to \code{TRUE}}

\item{token}{your authentication token. By default checks the environment variable \code{ARCGIS_TOKEN}}
}
\description{
Add Features to Feature Layer
}
