% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-mapping.R
\name{add_fields}
\alias{add_fields}
\title{fields will always take preference over .data}
\usage{
add_fields(
  feature,
  .data = NULL,
  fields = NULL,
  token = Sys.getenv("ARCGIS_TOKEN")
)
}
\description{
fields will always take preference over .data
}
