% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{add_offset}
\alias{add_offset}
\title{Add an offset to a query parameters}
\usage{
add_offset(offset, request, params)
}
\description{
\code{\link[=add_offset]{add_offset()}} takes a list of query parameters and creates a query request.
Importantly, this creates the paginated results that will be needed for
Feature Layers with more than 2000 observations.
}
\keyword{internal}
