% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{apply_filter_geom}
\alias{apply_filter_geom}
\title{Apply filter_geom to x}
\usage{
apply_filter_geom(
  x,
  filter_geom = NULL,
  predicate = "intersects",
  crs = sf::st_crs(x),
  error_call = rlang::caller_env()
)
}
\description{
Apply filter_geom to x
}
\keyword{internal}
