% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{arc_count}
\alias{arc_count}
\title{Count features matching a query}
\usage{
arc_count(
  x,
  ...,
  fields = NULL,
  where = NULL,
  crs = sf::st_crs(x),
  filter_geom = NULL,
  predicate = "intersects",
  token = arc_token()
)
}
\arguments{
\item{x}{an object of class \code{FeatureLayer}, \code{Table}, or \code{ImageServer}.}

\item{...}{additional query parameters passed to the API.}

\item{fields}{a character vector of the field names that you wish to be
returned. By default all fields are returned.}

\item{where}{a simple SQL where statement indicating which features should be
selected.}

\item{crs}{the spatial reference to be returned. If the CRS is different than
the CRS for the input \code{FeatureLayer}, a transformation will occur
server-side. Ignored if x is a \code{Table}.}

\item{filter_geom}{an object of class \code{bbox}, \code{sfc} or \code{sfg} used to filter
query results based on a predicate function.}

\item{predicate}{Spatial predicate to use with \code{filter_geom}. Default
\code{"intersects"}. Possible options are \code{"intersects"}, \code{"contains"},
\code{"crosses"},  \code{"overlaps"},  \code{"touches"}, and \code{"within"}.}

\item{token}{an \code{httr2_token} as created by \code{auth_code()} or similar}
}
\value{
A scalar integer
}
\description{
Returns the number of features a query would return without downloading them.
}
\examples{
\dontrun{
furl <- paste0(
  "https://services3.arcgis.com/ZvidGQkLaDJxRSJ2/arcgis/rest/services/",
  "PLACES_LocalData_for_BetterHealth/FeatureServer/0"
)

flayer <- arc_open(furl)

arc_count(flayer)
arc_count(flayer, where = "StateAbbr = 'RI'")
arc_count(flayer, fields = "StateAbbr", returnDistinctValues = "true")
}
}
