% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-open.R
\name{arc_open}
\alias{arc_open}
\title{Open connection to remote resource}
\usage{
arc_open(url, token = Sys.getenv("ARCGIS_TOKEN"))
}
\arguments{
\item{url}{The url of the remote resource. Must be of length one.}

\item{token}{}
}
\description{
Provided a URL, create an object referencing the remote resource.
Supports Feature Layers, Tables, Feature Servers, and Image Servers.
}
