% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-open.R
\name{arc_url_item}
\alias{arc_url_item}
\title{Convert a URL into a content URL}
\usage{
arc_url_item(
  id,
  key = "id",
  base_url = "https://www.arcgis.com/sharing/rest/content",
  path = "items",
  call = caller_env()
)
}
\description{
Convert a URL into a content URL
}
\keyword{internal}
