% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgislayers-package.R
\docType{package}
\name{arcgislayers-package}
\alias{arcgislayers}
\alias{arcgislayers-package}
\title{arcgislayers: An R Interface for ArcGIS REST Services}
\description{
An R package to interface with Esri's ArcGIS REST Services.
}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

Other contributors:
\itemize{
  \item Eli Pousson \email{eli.pousson@gmail.com} (\href{https://orcid.org/0000-0001-8280-1706}{ORCID}) [contributor]
  \item Kenneth Vernon \email{kenneth.b.vernon@gmail.com} (\href{https://orcid.org/0000-0003-0098-5092}{ORCID}) [contributor]
}

}
\keyword{internal}
