% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-attachment.R, R/attachments.R,
%   R/update-attachments.R
\name{add_attachments}
\alias{add_attachments}
\alias{query_layer_attachments}
\alias{download_attachments}
\alias{update_attachments}
\title{Query and Download Feature Service Attachments}
\usage{
add_attachments(
  x,
  feature_id,
  path,
  file_name = basename(path),
  .progress = TRUE,
  token = arc_token()
)

query_layer_attachments(
  x,
  definition_expression = "1=1",
  attachments_definition_expression = NULL,
  object_ids = NULL,
  global_ids = NULL,
  attachment_types = NULL,
  keywords = NULL,
  return_metadata = TRUE,
  ...,
  token = arc_token()
)

download_attachments(
  attachments,
  out_dir,
  ...,
  overwrite = FALSE,
  .progress = TRUE,
  token = arc_token()
)

update_attachments(
  x,
  feature_id,
  attachment_id,
  path,
  .progress = TRUE,
  token = arc_token()
)
}
\arguments{
\item{x}{an object of class \code{FeatureLayer}, \code{Table}, or \code{ImageServer}.}

\item{feature_id}{a vector of object IDs that corresponds to the feature of the corresponding \code{attachment_id}.}

\item{path}{a vecetor of the same length as \code{feature_id} indicating where the attachment exists.}

\item{file_name}{the name of the file. Defaults to the \code{basename(path)}.
Must be the same length as \code{feature_id}.}

\item{.progress}{default \code{TRUE.} Whether a progress bar should be provided.}

\item{token}{an \code{httr2_token} as created by \code{auth_code()} or similar}

\item{definition_expression}{default \code{1 = 1}. A SQL where clause that is applied to the layer. Only those records that conform to this expression will be returned. This parameter is required if neither \code{object_ids} or \code{global_ids} have been defined.}

\item{attachments_definition_expression}{default \code{NULL}. A SQL where calsue that is applied to the attachment metadata.
only attachments that conform to this expression will be returned.}

\item{object_ids}{mutually exclusive with \code{definition_expression} and \code{global_ids}. The object IDs of the features to query attachments of.}

\item{global_ids}{mutally exclusive with \code{definition_expression} and \code{object_ids}. The global IDs of the features to query attachments of.}

\item{attachment_types}{default \code{NULL}. A character vector of attachment types to filter on.}

\item{keywords}{default \code{NULL}. A character vector of the keywords to filter on.}

\item{return_metadata}{default \code{TRUE}. Returns metadata stored in the \code{exifInfo} field.}

\item{...}{unused}

\item{attachments}{a \code{data.frame} created by \code{query_layer_attachments()}. Must contain the columns \code{name}, \code{url}, and \code{contentType}.}

\item{out_dir}{the path to the folder to download the file}

\item{overwrite}{default \code{FALSE}. A}

\item{attachment_id}{the ID of the attachment—this corresponds to the \code{id} column returned from \code{query_layer_attachments()}}
}
\value{
\code{query_layer_attachments()} returns a data.frame.

\code{download_attachments()} returns a list. If an error occurs, the condition is captured and returned in the list.
Otherwise the path to the file that was downloaded is returned.

a \code{data.frame} with 2 columns returning the status of the update.
}
\description{
Query attachment information using \code{query_layer_attachments()} and
download attachments using \code{download_attachments()}.

Feature Services can contain attachments that are associated with a single feature ID.
\itemize{
\item Use \code{add_features()} to add attachments to a feature service
\item Use \code{update_features()} to update the attachments of a feature service
\item Use \code{query_layer_attachments()} to list attachments of a feature service
\item Use \code{download_attachments()} with the results of \code{query_layer_attachments()}
to download the attachments from a feature service locally
}
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
To rename or otherwise modify an attachment in a Feature Service, you must first download
that attachment, modify the file on disk, and then upload it again. This is a limitation
of ArcGIS Online and Enterprise. If you'd like to see this changed, please submit a community idea at \href{https://community.esri.com/t5/arcgis-online/ct-p/arcgis-online}{community.esri.com}.

If any requests fail, the requests are added as as the \code{errors} attribute to the resultant \code{data.frame}.
}
\examples{
\dontrun{
if (interactive()) {
library(arcgisutils)

# authenticate
set_arc_token(auth_user())

# open a feature service
feature_layer <- arc_open("your-item-id") |>
  # layer ID of the feature service
  get_layer(0)

# create a list of features to update
features <- c(1,2,3)

# create a list of files to upload as attachments
attachment_files <- c("path/to/file1.png", "path/to/file2.png", "path/to/file3.png")

# add the attachment files to the features in the feature layer
add_response <- add_attachments(feature_layer, features, attachment_files, use_basename=TRUE)
}
}
\dontrun{
# create a url path that isn't too wide for CRAN
furl <- paste(
  c(
    "https://services1.arcgis.com/hLJbHVT9ZrDIzK0I",
    "arcgis/rest/services/v8_Wide_Area_Search_Form_Feature_Layer___a2fe9c",
    "FeatureServer/0"
  ),
  collapse = "/"
)
# connect to the layer
layer <- arc_open(furl)

# get the attachment info
att <- query_layer_attachments(layer)

# download them to a path
download_attachments(att, "layer_attachments")
}
\dontrun{
if (interactive()) {
library(arcgisutils)

# authenticate
set_arc_token(auth_user())

# open a feature service
feature_layer <- arc_open("your-item-id") |>
  # layer ID of the feature service
  get_layer(0)

# query attachment layer information
attachments <- query_layer_attachments(feature_layer)

# create a temporary directory
tmp <- tempdir()

# download attachments to the temporary directory
download_attachments(attachments, tmp)

# get original paths
fps <- file.path(tmp, attachments$name)

# prepend attachments with the date
new_filenames <- paste0(Sys.Date(), "-", basename(attachments$name))

# create new file paths
new_fps <- file.path(dirname(fps), new_filenames)

# rename the files
file.rename(fps, new_fps)

# update the attachments
update_res <- update_attachments(
  feature_layer,
  # OID of the feature <> attachment relationship
  attachments$parentObjectId,
  # the attachment ID
  attachments$id,
  # the path to the attachment on disk
  new_fps
)
}
}
}
\references{
See \href{https://developers.arcgis.com/rest/services-reference/enterprise/add-attachment/#request-parameters}{API documentation} for more.

\href{https://developers.arcgis.com/rest/services-reference/enterprise/query-attachments-feature-service-layer/}{ArcGIS REST API Documentation}

See \href{https://developers.arcgis.com/rest/services-reference/enterprise/update-attachment/#request-parameters}{API documentation} for more.
}
