% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_null_or_scalar}
\alias{check_null_or_scalar}
\title{Useful for when an argument must either be NULL or a scalar
value. This is most useful when ensuring that values passed
to \code{httr2::req_body_multiform()} are scalars. Multiple length
values are not permitted.}
\usage{
check_null_or_scalar(
  x,
  arg = rlang::caller_arg(x),
  error_call = rlang::caller_env()
)
}
\description{
Useful for when an argument must either be NULL or a scalar
value. This is most useful when ensuring that values passed
to \code{httr2::req_body_multiform()} are scalars. Multiple length
values are not permitted.
}
\keyword{internal}
