% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_indices}
\alias{chunk_indices}
\title{Get chunk indices}
\usage{
chunk_indices(n, m)
}
\arguments{
\item{n}{the number of rows}

\item{m}{the chunk size}
}
\description{
For a given number of items and a chunk size, determine the start and end
positions of each chunk.
}
\keyword{internal}
