% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce_crs}
\alias{coalesce_crs}
\title{Pick first non-missing CRS}
\usage{
coalesce_crs(x, y)
}
\arguments{
\item{x}{an object of class \code{crs}}

\item{y}{an object of class \code{crs}}
}
\description{
Pick first non-missing CRS
}
\examples{

x <- sf::st_crs(27572)
y <- sf::st_crs(NA)

coalesce_crs(x, y)
}
\keyword{internal}
