% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-layer-filter.R
\name{collect.FeatureLayer}
\alias{collect.FeatureLayer}
\title{Collect}
\usage{
\method{collect}{FeatureLayer}(x, geometry = TRUE, token = "", ...)
}
\description{
Collect
}
\details{
The Feature Layer method of `collect()` will overwrite the `returnGeometry` parameter if set with `update_params()`. Use the `geometry` argument in `collect()` to set the parameter.
}
