% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{collect_layer}
\alias{collect_layer}
\title{Query a FeatureLayer or Table object}
\usage{
collect_layer(
  x,
  n_max = Inf,
  token = Sys.getenv("ARCGIS_TOKEN"),
  ...,
  error_call = rlang::caller_env()
)
}
\description{
\code{\link[=collect_layer]{collect_layer()}} is the "workhorse" function that actually executes the
queries for FeatureLayer or Table objects.
}
\keyword{internal}
