% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_features}
\alias{count_features}
\title{Count the number of features in a feature layer}
\usage{
count_features(request, token)
}
\arguments{
\item{request}{the base request created from the feature layer url.}
}
\value{
a numeric with the total number of features in the feature layer
}
\description{
Count the number of features in a feature layer
}
\keyword{internal}
