% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{count_results}
\alias{count_results}
\title{Given a query, determine how many features will be returned}
\usage{
count_results(req, query, token)
}
\description{
Given a query, determine how many features will be returned
}
