% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_features.R
\name{delete_features}
\alias{delete_features}
\title{Delete Features from Feature Layer}
\usage{
delete_features(
  x,
  object_ids,
  where,
  filter_geom,
  predicate = "intersects",
  rollback_on_fail = TRUE,
  token = Sys.getenv("ARCGIS_TOKEN"),
  ...
)
}
\arguments{
\item{x}{a \code{FeatureLayer}}

\item{object_ids}{a numeric vector of object IDs to be deleted.}

\item{where}{a simple SQL where statement indicating which features should be deleted.
When the where statement evaluates to \code{TRUE}, those values will be deleted.}

\item{filter_geom}{an \code{sfc} or \code{sfg} object. If \code{sfc} its length must be one.}

\item{predicate}{deault \code{"intersects"}. The spatial predicate to be used for the spatial filter. Ignored if \code{filter_geom} is not provided.}

\item{rollback_on_fail}{default \code{TRUE}. Specifies whether the edits should be applied only if all submitted edits succeed.}

\item{token}{your authorization token. By default uses the environment variable \code{ARCGIS_TOKEN} as set by \code{set_auth_token()}.}
}
\description{
Delete features from a feature layer based on object ID, a where clause, or a spatial filter.
}
