% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_errors}
\alias{detect_errors}
\title{Detect errors in parsed json response}
\usage{
detect_errors(response)
}
\description{
The requests responses from ArcGIS don't return the status code
in the response itself but rather from the body in the json.
So this function checks for the existence of the error field.
}
\keyword{internal}
