% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_geometry.R
\name{determine_dims}
\alias{determine_dims}
\alias{determine_dims.sfc}
\alias{determine_dims.sfg}
\title{Determine the dimensions of a geometry object}
\usage{
determine_dims(x)

\method{determine_dims}{sfc}(x)

\method{determine_dims}{sfg}(x)
}
\value{
a scalar character of the value \code{"xy"}, \code{"xyz"}, or \code{"xyzm"} depending
on what dimensions are represented.
}
\description{
Given an sfc or sfg object determine what dimensions are represented.
}
\keyword{internal}
