% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-layer-filter.R
\name{select.FeatureLayer}
\alias{select.FeatureLayer}
\alias{select.Table}
\alias{filter.FeatureLayer}
\alias{filter.Table}
\alias{collect.FeatureLayer}
\alias{dplyr}
\title{dplyr methods}
\usage{
\method{select}{FeatureLayer}(x, ...)

\method{select}{Table}(x, ...)

\method{filter}{FeatureLayer}(x, ...)

\method{filter}{Table}(x, ...)

\method{collect}{FeatureLayer}(x, n_max = Inf, token = Sys.getenv("ARCGIS_TOKEN"), ...)
}
\description{
dplyr methods
}
\details{
The Feature Layer method of `collect()` will overwrite the `returnGeometry` parameter if set with `update_params()`. Use the `geometry` argument in `collect()` to set the parameter.
}
