% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-constructors-legacy.R
\name{feature_server}
\alias{feature_server}
\title{Create a feature server}
\usage{
feature_server(url, token = Sys.getenv("ARCGIS_TOKEN"))
}
\arguments{
\item{url}{the url of a feature server}
}
\description{
Create a feature server
}
\examples{
url <- "https://services2.arcgis.com/j80Jz20at6Bi0thr/ArcGIS/rest/services/hexagons_state/FeatureServer"
feature_server(url)
}
\keyword{internal}
