% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{feature_table}
\alias{feature_table}
\title{Feature Table}
\usage{
feature_table(url, token = "")
}
\description{
Feature Table
}
\examples{
tbl_url <- "https://services2.arcgis.com/j80Jz20at6Bi0thr/ArcGIS/rest/services/List_of_Providers/FeatureServer/27"

feature_table(tbl_url)
}
