% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_featureset.R
\name{featureset_geometry}
\alias{featureset_geometry}
\title{Convert an object to featureset list structure}
\usage{
featureset_geometry(x)
}
\arguments{
\item{x}{an object of class \code{sfc} or \code{sf}}
}
\description{
The output of this is intended to be passed to \code{jsonify::to_json()}
}
\keyword{internal}
