% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_layer_metadata}
\alias{fetch_layer_metadata}
\title{Retrieve metadata for a feature layer}
\usage{
fetch_layer_metadata(request, token)
}
\arguments{
\item{request}{an httr2 request object. Should be the \code{base_req}
object that is created from the provided feature layer url}
}
\description{
Retrieve metadata for a feature layer
}
\keyword{internal}
