% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri-field-mapping.R
\name{remote_ptype_tbl}
\alias{remote_ptype_tbl}
\alias{get_ptype}
\title{Create a lazy frame prototype}
\usage{
remote_ptype_tbl(fields)

get_ptype(field_type)
}
\description{
Given the fields of a feature layer create a lazy frame with
the name field names and the corresponding R type. Used for partial_eval
}
\details{
\itemize{
\item \code{get_ptype()} takes a scalar character containing the Esri field type and returns a prototype of the pertinent R type
\item \code{remote_ptype_tbl()} takes a data frame of fields as derived from \code{list_fields()} and creates a lazy table
}
}
\keyword{internal}
