% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-layer-filter.R
\name{filter}
\alias{filter}
\alias{filter.FeatureLayer}
\alias{filter.Table}
\title{Filter methods}
\usage{
filter(x, ...)

\method{filter}{FeatureLayer}(x, ...)

\method{filter}{Table}(x, ...)
}
\description{
Filter methods
}
