% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-server.R
\name{get_feature}
\alias{get_feature}
\alias{get_all_features}
\title{Extract a feature layer}
\usage{
get_feature(x, id = 0, token = "")

get_all_features(x, token = "")
}
\value{
`get_feature()` returns a single `FeatureLayer` based on its ID
`get_all_features()` returns a named `list` with an element `layers` and `tables`. Each a list containing `FeatureLayer` and `Table`s respectively.
}
\description{
From a Feature Server, extract a feature layer.
}
