% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-server.R, R/get-layer.R
\name{get_layer}
\alias{get_layer}
\alias{get_all_layers}
\title{Extract a feature layer}
\usage{
get_layer(x, fields, where, crs = st_crs(x), n_max = Inf, ...)

get_all_layers(x, token = "")

get_layer(x, fields, where, crs = st_crs(x), n_max = Inf, ...)
}
\arguments{
\item{x}{an object of class `FeatureLayer` or `Table`.}

\item{fields}{a character vector of the field names that you wish to be returned. By default all fields are returned.}

\item{crs}{the spatial reference to be returned. If the CRS is different than the `FeatureLayer`'s CRS, a transformation will occur server-side. Ignored for `Table` objects.}

\item{n_max}{the maximum number of features to return. By default returns every feature available.}

\item{...}{additional query parameters passed to `update_params()`. See [reference documentation](https://developers.arcgis.com/rest/services-reference/enterprise/query-feature-service-layer-.htm#GUID-BC2AD141-3386-49FB-AA09-FF341145F614) for possible arguments.}
}
\value{
`get_layer()` returns a single `FeatureLayer` or `Table` based on its ID
`get_all_layers()` returns a named `list` with an element `layers` and `tables`. Each a list containing `FeatureLayer` and `Table`s respectively.
}
\description{
From a Feature Server, extract a feature layer.

Give a `FeatureLayer` or `Table` object, retrieve its data as an `sf` object or `tibble` resepctively.
}
