% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-layer-helpers.R
\name{head.FeatureLayer}
\alias{head.FeatureLayer}
\title{Fetch n rows from a feature layer}
\usage{
\method{head}{FeatureLayer}(x, n = 6, token = "", ...)
}
\description{
Fetch n rows from a feature layer
}
