% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer-constructors-legacy.R
\name{image_server}
\alias{image_server}
\title{Image Server Representation}
\usage{
image_server(url, token = Sys.getenv("ARCGIS_TOKEN"))
}
\description{
Image Server Representation
}
\keyword{internal}
