% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esri-field-mapping.R
\name{infer_esri_type}
\alias{infer_esri_type}
\title{Given a data frame infer the esri field types}
\usage{
infer_esri_type(.data)
}
\description{
Given a data frame infer the esri field types
}
\keyword{internal}
