% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_fields}
\alias{list_fields}
\title{List fields in a a feature layer}
\usage{
list_fields(x)
}
\arguments{
\item{x}{and object of class \code{FeatureLayer}, \code{Table}, or \code{ImageServer}.}
}
\description{
List fields in a a feature layer
}
