% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-feature-layer.R
\name{make_offset_reqs}
\alias{make_offset_reqs}
\title{Create paginated requests}
\usage{
make_offset_reqs(request, offsets)
}
\arguments{
\item{request}{the base request created from only the feature layer
url.}

\item{offsets}{a vector of offsets. Used to create multiple requests
for each requisite page.}
}
\description{
For a feature layer, create a list of requests that use the
appropriate offsets. This allows us to not paginate but rather
create the requests in parallel with \code{httr2::multi_req_perform}.
}
\keyword{internal}
