% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geos-to-sfc-compatibility.R
\name{make_sf_compat}
\alias{make_sf_compat}
\title{Cast geos geometry to sfc object}
\usage{
make_sf_compat(geometry)
}
\arguments{
\item{geom_types}{}
}
\description{
geos geometry cannot be used directly by sf so it must be converted. Conversion
using sf is slow. This function checks the types of geometries from our geojson
it then casts them "up" (e.g. from polygon to multipolygon) if necessary and
sets the propery geometry type for sf. Then casts to sf. This saves a lot of time.
}
\details{
Based on https://github.com/paleolimbot/geos/issues/80.
}
\keyword{internal}
