% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{obj_check_arcgis}
\alias{obj_check_arcgis}
\title{Check if x is a FeatureLayer or Table object}
\usage{
obj_check_arcgis(
  x,
  class = c("FeatureLayer", "Table"),
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{An object to test for inheritance.}

\item{class}{A character vector of classes.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Check if x is a FeatureLayer or Table object
}
\keyword{internal}
