% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{obj_check_layer}
\alias{obj_check_layer}
\title{Check if x is a FeatureLayer or Table class object}
\usage{
obj_check_layer(
  x,
  class = c("FeatureLayer", "Table"),
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\description{
Check if x is a FeatureLayer or Table class object
}
\keyword{internal}
