% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_spatial_filter}
\alias{prepare_spatial_filter}
\title{Prepare json for spatial filters}
\usage{
prepare_spatial_filter(x, y, predicate)
}
\arguments{
\item{x}{is a \code{FeatureLayer}}

\item{y}{is an \code{sf} object}
}
\description{
Prepare json for spatial filters
}
\keyword{internal}
