% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-layer-class.R
\name{prettify_param}
\alias{prettify_param}
\title{function to make printing easier}
\usage{
prettify_param(param, query)
}
\description{
The function takes a parameter name and the query list and trims each element
to fit on the contents of the window.
}
\keyword{internal}
