% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-layer.R
\name{query_layer}
\alias{query_layer}
\title{Retrieve a feature layer}
\usage{
query_layer(
  x,
  fields,
  where,
  crs = st_crs(x),
  filter_geom,
  predicate = "intersects",
  n_max = Inf,
  ...
)
}
\arguments{
\item{x}{an object of class \code{FeatureLayer} or \code{Table}.}

\item{fields}{a character vector of the field names that you wish to be returned. By default all fields are returned.}

\item{crs}{the spatial reference to be returned. If the CRS is different than the \code{FeatureLayer}'s CRS, a transformation will occur server-side. Ignored for \code{Table} objects.}

\item{n_max}{the maximum number of features to return. By default returns every feature available.}

\item{...}{additional query parameters passed to \code{update_params()}. See \href{https://developers.arcgis.com/rest/services-reference/enterprise/query-feature-service-layer-.htm#GUID-BC2AD141-3386-49FB-AA09-FF341145F614}{reference documentation} for possible arguments.}
}
\description{
Give a \code{FeatureLayer} or \code{Table} object, retrieve its data as an \code{sf} object or \code{tibble} resepctively.
}
