% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-read.R
\name{set_layer_col_names}
\alias{set_layer_col_names}
\title{Set and repair column names for FeatureLayer or Table data frame}
\usage{
set_layer_col_names(
  .data,
  .layer = NULL,
  col_names = TRUE,
  name_repair = NULL,
  alias = c("drop", "label", "replace"),
  call = rlang::caller_env()
)
}
\arguments{
\item{.data}{A data frame returned by \code{arc_select() }or \code{arc_read()}.}

\item{.layer}{A Table or FeatureLayer object. Required if \code{alias} is \code{"label"} or
\code{"replace"}.}

\item{col_names}{Default \code{TRUE}. Column names or name handling rule.
\code{col_names} can be \code{TRUE}, \code{FALSE}, \code{NULL}, or a character vector:
\itemize{
\item If \code{TRUE}, use existing default column names for the layer or table.
If \code{FALSE} or \code{NULL}, column names will be generated automatically: X1, X2,
X3 etc.
\item If \code{col_names} is a character vector, values replace the existing column
names. \code{col_names} can't be length 0 or longer than the number of fields in
the returned layer.
}}

\item{name_repair}{Default \code{"unique"}. See \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}} for
details. If \code{name_repair = NULL}, names are set directly.}

\item{alias}{Use of field alias values. Default \verb{c("drop", "label", "replace"),}. There are three options:
\itemize{
\item \code{"drop"}, field alias values are ignored.
\item \code{"label"}: field alias values are assigned as a label attribute for each field.
\item \code{"replace"}: field alias values replace existing column names. \code{col_names}
must \code{TRUE} for this option to be applied.
}}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=set_layer_col_names]{set_layer_col_names()}} can replace or label column names based on the the
field aliases from a corresponding \code{Table} or \code{FeatureLayer} object created
with \code{arc_open()}. Optionally repair names using \code{\link[vctrs:vec_as_names]{vctrs::vec_as_names()}}.
}
