% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_features.R
\name{st_as_features}
\alias{st_as_features}
\alias{st_as_features.sfc}
\alias{st_as_features.sf}
\alias{st_as_features.data.frame}
\title{Create Esri Feature JSON array
Represent simple features as Esri JSON features.}
\usage{
st_as_features(x, ...)

\method{st_as_features}{sfc}(x, ...)

\method{st_as_features}{sf}(x, ...)

\method{st_as_features}{data.frame}(x, ...)
}
\arguments{
\item{x}{an sf or sfc class object}

\item{...}{unused}
}
\description{
Create Esri Feature JSON array
Represent simple features as Esri JSON features.
}
\examples{

# create sfc object of points from multipoint
mpnt <- st_multipoint(
  matrix(runif(10, min = -180, max = 180), ncol = 2)
)

pnt_sfc <- st_cast(st_sfc(mpnt), "POINT")

# sfc method
st_as_features(pnt_sfc)

# sf method
pnt_sf <- st_sf(pnt_sfc)
pnt_sf[["id"]] <- 1:5

st_as_features(pnt_sf)
}
