% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_featureset.R
\name{st_as_featureset}
\alias{st_as_featureset}
\alias{st_as_featureset.sfc}
\alias{st_as_featureset.sf}
\alias{st_as_featureset.data.frame}
\title{Create Esri FeatureSet JSON}
\usage{
st_as_featureset(x, ...)

\method{st_as_featureset}{sfc}(x, crs = 4326, ...)

\method{st_as_featureset}{sf}(x, crs = sf::st_crs(x), ...)

\method{st_as_featureset}{data.frame}(x, ...)
}
\arguments{
\item{x}{an sf or sfc class object}

\item{...}{unused}

\item{crs}{the coordinate reference system of the FeatureSet. Must be interpretable by `sf::st_crs()`}
}
\description{
Represent simple features as Esri JSON.
}
\examples{
# create sfc object of points from multipoint
mpnt <- st_multipoint(
  matrix(runif(10, min = -180, max = 180), ncol = 2)
)

pnt_sfc <- st_cast(st_sfc(mpnt), "POINT")

# sfc method
st_as_featureset(pnt_sfc, crs = 4326)

# sf method
pnt_sf <- st_sf(pnt_sfc)
pnt_sf[["id"]] <- 1:5

st_as_featureset(pnt_sf, crs = 4326)
}
