% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_geometry.R
\name{st_as_geometry}
\alias{st_as_geometry}
\alias{st_as_geometry.POINT}
\alias{st_as_geometry.MULTIPOINT}
\alias{st_as_geometry.LINESTRING}
\alias{st_as_geometry.MULTILINESTRING}
\alias{st_as_geometry.POLYGON}
\alias{st_as_geometry.MULTIPOLYGON}
\alias{st_as_geometry.envelope}
\title{Convert an sf object to esri geometry}
\usage{
st_as_geometry(x, crs, ...)

\method{st_as_geometry}{POINT}(x, crs = 4326, ...)

\method{st_as_geometry}{MULTIPOINT}(x, crs = 4326, ...)

\method{st_as_geometry}{LINESTRING}(x, crs = 4326, ...)

\method{st_as_geometry}{MULTILINESTRING}(x, crs = 4326, ...)

\method{st_as_geometry}{POLYGON}(x, crs = 4326, ...)

\method{st_as_geometry}{MULTIPOLYGON}(x, crs = 4326, ...)

\method{st_as_geometry}{envelope}(x, crs = sf::st_crs(x))
}
\arguments{
\item{x}{an object of class `sfg`}

\item{crs}{a CRS ID, crs object, or a well-known text representation of CRS}
}
\description{
Convert an sf object to esri geometry
}
\examples{
library(sf)
st_as_geometry(st_point(c(0, 1, 3, 4)))
st_as_geometry(st_multipoint(x = matrix(runif(4), ncol = 2)))
st_as_geometry(st_linestring(x = matrix(runif(2), ncol = 2)))
st_as_geometry(st_linestring(x = matrix(runif(2), ncol = 2)))

# polygon
m <- matrix(
  c(0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 2, 2, 1, 2, 3, 1, 3, 2, 0, 0, 0),
  ncol = 3,
  byrow = TRUE
)
st_as_geometry(st_polygon(list(m)))
}
