% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_as_features.R, R/st_as_featureset.R,
%   R/st_as_geometry.R
\name{st_as_features}
\alias{st_as_features}
\alias{st_as_featureset}
\alias{st_as_json}
\title{Create Esri FeatureSet JSON}
\usage{
st_as_features(x, ...)

st_as_featureset(x, ...)

st_as_json(x, crs, ...)
}
\arguments{
\item{x}{an object of class \code{sfg}}

\item{...}{unused}

\item{crs}{a CRS ID, crs object, or a well-known text representation of CRS}
}
\description{
Represent simple features as Esri JSON.

These functions convert R objects to Esri json representations. There are three
types of representations. These are, from smallest to largest, a \href{https://developers.arcgis.com/documentation/common-data-types/geometry-objects.htm}{geometry object}, a \href{https://developers.arcgis.com/documentation/common-data-types/feature-object.htm}{feature} and a \href{https://developers.arcgis.com/documentation/common-data-types/featureset-object.htm}{feature set}.
}
\details{
\itemize{
\item \code{st_as_json()} converts an \code{sfg} object to a geometry object
\item \code{st_as_features()} converts an \code{sfc} or \code{sf} object to a list of features
\item \code{st_as_featureset()} converts an \code{sf}, \code{sfc}, or \code{data.frame} to a feature set object
}

Geometry object contain the coordinates for a geometry. Features are geometries that
have associated attributes with them. This would be similar to a row in an sf object.
FeatureSets are a list of features that have additional metadata fields such as
\code{spatialReference}, \code{geomtryType}, and \code{fields}. FeatureSets correspond to an
\code{sf} object.

Geometry objects are defined for 5 different types. These are:
\itemize{
\item Point: \code{esriGeometryPoint}
\item Multipoint: \code{esriGeometryMultipoint}
\item Polyline: \code{esriGeometryPolyline}
\itemize{
\item note that polyline encompasses both LINESTRING and MULTILINESTRING
}
\item Polygon: \code{esriGeometryPolygon}
\itemize{
\item note that polygon encompasses both POLYGON and MULTIPOLYGON
}
\item Envelope: \code{esriGeometryEnvelope}
\itemize{
\item envelopes correspond with bounding boxes but can have a Z dimension
}
}
}
\examples{
library(sf)
st_as_json(st_point(c(0, 1, 3, 4)))
st_as_json(st_multipoint(x = matrix(runif(4), ncol = 2)))
st_as_json(st_linestring(x = matrix(runif(2), ncol = 2)))
st_as_json(st_linestring(x = matrix(runif(2), ncol = 2)))

# polygon
m <- matrix(
  c(0, 0, 0, 0, 0, 1, 0, 1, 1, 1, 2, 2, 1, 2, 3, 1, 3, 2, 0, 0, 0),
  ncol = 3,
  byrow = TRUE
)
st_as_json(st_polygon(list(m)))
}
