% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_envelope.R
\name{st_envelope}
\alias{st_envelope}
\alias{st_envelope.sfc}
\alias{st_envelope.sf}
\alias{st_envelope.sfg}
\alias{new_envelope}
\alias{print.envelope}
\title{Return an envelope of a simple feature or simple feature set}
\usage{
st_envelope(x, crs)

\method{st_envelope}{sfc}(x, crs = sf::st_crs(x))

\method{st_envelope}{sf}(x, crs = sf::st_crs(x))

\method{st_envelope}{sfg}(x, crs = 4326)

new_envelope(x, crs)

\method{print}{envelope}(x, digits = 4, ...)
}
\arguments{
\item{x}{an object of class \code{sfg}, \code{sfc}, or \code{sf}}
}
\description{
Return an envelope of a simple feature or simple feature set
}
\details{
An envelope is much like a bounding box. In fact, it is a bounding box when data
is only two dimensional. An envelope records the mimimum and maximum's of each
dimension. In 2D space this is \code{xmin}, \code{ymin}, \code{xmax}, and \code{ymax}. When a Z or M
dimension is present, it also records the min and maxes of those as well.
}
\examples{
st_envelope(sf::st_point())
}
