% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_definition.R
\name{update_definition}
\alias{update_definition}
\title{Update a Feature Layer or Feature Service definition}
\usage{
update_definition(x = NULL, ..., async = FALSE, token = arc_token())
}
\arguments{
\item{x}{A Feature Layer or Feature Service class object.}

\item{...}{Additional parameters "updateDefinition" body for the request.}

\item{async}{Default \code{FALSE}. If \code{TRUE}, support asynchronous processing for
the request.}

\item{token}{your authorization token.}
}
\description{
\code{\link[=update_definition]{update_definition()}} supports updating a definition property for a hosted
Feature Service or Feature Layer. Examples of properties include the layer
name, renderer, or field properties. Parameters passed to \code{...} must have
names matching the definitions. Parameters are converted to a JSON
\code{updateDefinition} parameter using \code{\link[jsonify:to_json]{jsonify::to_json()}}. See the ArcGIS REST API documentation on Administer Hosted Feature Services
for more details on updating the definition for \href{https://developers.arcgis.com/rest/services-reference/online/update-definition-feature-service-.htm}{a FeatureService}
or \href{https://developers.arcgis.com/rest/services-reference/online/update-definition-feature-layer/}{a FeatureLayer}.
}
