% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_features.R
\name{update_features}
\alias{update_features}
\title{Update Feature Layer}
\usage{
update_features(
  x,
  .data,
  token = Sys.getenv("ARCGIS_TOKEN"),
  rollback_on_failure = TRUE,
  ...
)
}
\description{
Update Feature Layer
}
