% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_layer.R
\name{update_layer}
\alias{update_layer}
\alias{.update_layer_name}
\alias{.update_layer_properties}
\title{Update Feature Layer definition}
\usage{
update_layer(x, ..., update = NULL, async = FALSE, token = arc_token())

.update_layer_name(...)

.update_layer_properties(
  description = NULL,
  license_text = NULL,
  xss_trusted_fields = NULL,
  display_field = NULL,
  min_scale = NULL,
  max_scale = NULL,
  max_records = NULL,
  default_visibility = NULL,
  ...
)
}
\description{
\url{https://developers.arcgis.com/rest/services-reference/online/update-definition-feature-layer-.htm#ESRI_SECTION2_027B8A04953C409E9A9D9051DD4CDBEE}
}
