% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{update_params}
\alias{update_params}
\title{Modify query parameters}
\usage{
update_params(x, ...)
}
\arguments{
\item{x}{a \code{FeatureLayer} or \code{Table} object}

\item{...}{key value pairs of query parameters and values.}
}
\description{
\code{\link[=update_params]{update_params()}} takes named arguments and updates the query.
}
\keyword{internal}
