% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_query}
\alias{clear_query}
\alias{list_fields}
\alias{refresh_layer}
\title{Clear all query parameters}
\usage{
clear_query(x)

list_fields(x)

refresh_layer(x)
}
\arguments{
\item{x}{and object of class \code{FeatureLayer}, \code{Table}, or \code{ImageServer}.}
}
\description{
Useful to update metadata after modifying a remote.
}
