% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{validate_params}
\alias{validate_params}
\title{Validate query parameters}
\usage{
validate_params(params, token)
}
\description{
\code{\link[=validate_params]{validate_params()}} ensures that the parameters are set to minimally
acceptable values.
}
\keyword{internal}
