% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-layer.R
\name{add_offset}
\alias{add_offset}
\title{This function takes a list of query parameters and creates a query request
Importantly, this creates the paginated results that will be needed for
Feature Layers with more than 2000 observations}
\usage{
add_offset(offset, request, params)
}
\description{
This function takes a list of query parameters and creates a query request
Importantly, this creates the paginated results that will be needed for
Feature Layers with more than 2000 observations
}
\keyword{internal}
