% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-open.R
\name{arc_open}
\alias{arc_open}
\title{Open connection to remote resource}
\usage{
arc_open(url, token = Sys.getenv("ARCGIS_TOKEN"))
}
\arguments{
\item{url}{The url of the remote resource. Must be of length one.}

\item{token}{your authorization token. By default checks the environment
variable \code{ARCGIS_TOKEN}. Set your token using \code{arcgisutils::set_auth_token()}.}
}
\value{
Depending on the provided URL returns a \code{FeatureLayer}, \code{Table}, \code{FeatureServer}, or \code{ImageServer}.
}
\description{
Provided a URL, create an object referencing the remote resource.
The resultant object acts as a reference to the remote data source.
}
\details{
To extract data from the remote resource utilize \code{\link[=arc_select]{arc_select()}} for objects of class
\code{FeatureLayer} or \code{Table}. For \code{ImageServer}s, utilize \code{\link[=arc_raster]{arc_raster()}}.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
arc_select arc_raster
}
