% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-raster.R
\name{arc_raster}
\alias{arc_raster}
\title{Retrieve Imagery}
\usage{
arc_raster(
  x,
  xmin,
  xmax,
  ymin,
  ymax,
  crs = sf::st_crs(x),
  width = NULL,
  height = NULL,
  format = "tiff",
  token = Sys.getenv("ARCGIS_TOKEN")
)
}
\arguments{
\item{x}{an \code{ImageServer} as created with \code{image_server()}.}

\item{xmin}{the minimum bounding longitude value.}

\item{xmax}{the maxmimum bounding longitude value.}

\item{ymin}{that minimum bounding latitude value.}

\item{ymax}{the maximum bounding latitude value.}

\item{crs}{the CRS of the resultant raster image and the provided bounding box defined by \code{xmin}, \code{xmax}, \code{ymin}, \code{ymax} (passed to \code{bboxSR} and \code{outSR} query parameters).}

\item{width}{default \code{NULL}. Cannot exceed \code{x[["maxImageWidth"]]}.}

\item{height}{default \code{NULL}. Cannot exceed \code{x[["maxImageHeight"]]}.}

\item{format}{default \code{"tiff"}. Must be one of "jpgpng", "png", "png8", "png24", "jpg", "bmp", "gif", "tiff", "png32", "bip", "bsq", "lerc".}

\item{token}{authorization token fetched from the environment variable \code{ARCGIS_TOKEN}}
}
\value{
An object of class \code{SpatRaster}.
}
\description{
Retrieve Imagery
}
\examples{
if (interactive()) {
img_url <- "https://landsat2.arcgis.com/arcgis/rest/services/Landsat/MS/ImageServer"

landsat <- arc_open(img_url, token = "")

bbox <- sf::st_bbox(c(xmin = -71, ymin = 43, xmax = -67, ymax = 47.5), crs = 4326)

arc_raster(landsat, bbox, 1000, 1000)

}

}
