% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_client}
\alias{auth_client}
\alias{auth_code}
\alias{set_auth_token}
\title{Authentication}
\usage{
auth_client(
  client = Sys.getenv("ARCGIS_CLIENT"),
  secret = Sys.getenv("ARCGIS_SECRET"),
  host = "https://www.arcgis.com"
)

auth_code(
  client = Sys.getenv("ARCGIS_CLIENT"),
  host = "https://www.arcgis.com"
)

set_auth_token(token, quietly = FALSE)
}
\description{
Authentication
}
