% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_code}
\alias{auth_code}
\alias{auth_client}
\alias{auth_user}
\alias{set_auth_token}
\alias{refresh_token}
\alias{validate_or_refresh_token}
\title{Authorization}
\usage{
auth_code(
  client = Sys.getenv("ARCGIS_CLIENT"),
  host = "https://www.arcgis.com"
)

auth_client(
  client = Sys.getenv("ARCGIS_CLIENT"),
  secret = Sys.getenv("ARCGIS_SECRET"),
  host = "https://www.arcgis.com",
  expiration = 120
)

auth_user(
  username = Sys.getenv("ARCGIS_USER"),
  password = Sys.getenv("ARCGIS_PASSWORD"),
  host = "https://www.arcgis.com",
  expiration = 60
)

set_auth_token(token, quietly = FALSE)

refresh_token(
  token,
  client = Sys.getenv("ARCGIS_CLIENT"),
  host = "https://arcgis.com"
)

validate_or_refresh_token(
  token,
  client = Sys.getenv("ARCGIS_CLIENT"),
  host = "https://arcgis.com",
  refresh_threshold = 0
)
}
\arguments{
\item{client}{an OAuth 2.0 developer application client ID. By default uses the
environment variable \code{ARCGIS_CLIENT}.}

\item{host}{default \verb{"https://www.arcgis.com}}

\item{secret}{an OAuth 2.0 developer application secret. By default uses the environment
variable \code{ARCGIS_SECRET}.}

\item{expiration}{the duration of the token in minutes.}
}
\description{
Authorize your R session to connect to an ArcGIS Portal. ArcGIS Online and ArcGIS Enterprise
utilize OAuth2 authorization flow.
}
\details{
Create a OAuth2.0 ArcGIS Application at https://developers.arcgis.com/applications/
\itemize{
\item \code{set_auth_token()} is a helper function that takes an \code{httr2_token} as created by \code{auth_code()} or \code{auth_client()} and sets the \code{ARCGIS_TOKEN} environment variable
}
}
