% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_code}
\alias{auth_code}
\alias{auth_client}
\alias{auth_password}
\alias{set_auth_token}
\alias{refresh_token}
\alias{validate_or_refresh_token}
\title{Authenticate to a remote resource}
\usage{
auth_code(
  client = Sys.getenv("ARCGIS_CLIENT"),
  host = "https://www.arcgis.com"
)

auth_client(
  client = Sys.getenv("ARCGIS_CLIENT"),
  secret = Sys.getenv("ARCGIS_SECRET"),
  host = "https://www.arcgis.com",
  expiration = 120
)

auth_password(
  username = Sys.getenv("ARCGIS_USER"),
  password = Sys.getenv("ARCGIS_PASSWORD"),
  host = "https://www.arcgis.com",
  expiration = 60
)

set_auth_token(token, quietly = FALSE)

refresh_token(
  client = Sys.getenv("ARCGIS_CLIENT"),
  host = "https://arcgis.com",
  token
)

validate_or_refresh_token(
  token,
  client = Sys.getenv("ARCGIS_CLIENT"),
  host = "https://arcgis.com",
  refresh_threshold = 0
)
}
\description{
In order to access secured content one must authenticate to your remote resource—
either ArcGIS Online (default) or ArcGIS enterprise. These functions perform
Oauth2 authentication with the remote resource and return you a token
}
