% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{check_inherits_any}
\alias{check_inherits_any}
\title{Check if x inherits any of the supplied class values and error if not}
\usage{
check_inherits_any(
  x,
  class,
  arg = rlang::caller_arg(x),
  call = rlang::caller_env()
)
}
\arguments{
\item{x}{Vector that will be collapsed by cli.}

\item{class}{A character vector of classes.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Check if x inherits any of the supplied class values and error if not
}
\keyword{internal}
