% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-mapping.R
\name{is_date}
\alias{is_date}
\alias{date_to_ms}
\title{Date handling}
\usage{
is_date(x)

date_to_ms(x, tz = "UTC")
}
\description{
Date handling
}
\details{
- `is_date()` checks to see if an object is a date
- `date_to_ms()` converts a date to miliseconds from Unix Epoch
}
\keyword{internal}
