% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-definition.R
\name{add_definition}
\alias{add_definition}
\alias{update_definition}
\alias{delete_definition}
\title{Add, update, or delete a Feature Layer, Table, or Feature Service definition}
\usage{
add_definition(x, ..., async = FALSE, token = arc_token())

update_definition(x, ..., async = FALSE, token = arc_token())

delete_definition(x, ..., async = FALSE, token = arc_token())
}
\arguments{
\item{x}{A Feature Layer, Table, or Feature Service class object.}

\item{...}{Additional parameters for the "addToDefinition" or "updateDefinition" body of the request.}

\item{async}{Default \code{FALSE}. If \code{TRUE}, support asynchronous processing for
the request.}

\item{token}{your authorization token.}
}
\value{
If \code{async = FALSE}, return an updated "FeatureServer" or "FeatureLayer" object with the added, updated, or deleted definitions. If \code{async = TRUE}, the input Feature Layer or Feature Server object \code{x} is returned as is.
}
\description{
\code{\link[=add_definition]{add_definition()}} and \code{\link[=update_definition]{update_definition()}} support adding or updating
definition properties for a hosted Feature Service or Feature Layer.
\code{\link[=delete_definition]{delete_definition()}} supports deleting a definition. Examples of properties
include the layer name, renderer, or field properties. Named parameters
passed to \code{...} must have names matching supported definitions. Parameters
are converted to a JSON \code{addToDefinition}, \code{updateDefinition}, or
\code{deleteFromDefinition} query parameter using \code{\link[jsonify:to_json]{jsonify::to_json()}}.
}
\details{
See the ArcGIS REST API documentation on Administer Hosted Feature Services
for more details:
\itemize{
\item see the \href{https://developers.arcgis.com/web-map-specification/objects/layerDefinition/}{layerDefinition} object documentation.
\item adding definitions for a \href{https://developers.arcgis.com/rest/services-reference/online/add-to-definition-feature-layer/}{FeatureLayer} or \href{https://developers.arcgis.com/rest/services-reference/online/add-to-definition-feature-service/}{a FeatureService}
\item updating definitions for \href{https://developers.arcgis.com/rest/services-reference/online/update-definition-feature-layer/}{a FeatureLayer} or \href{https://developers.arcgis.com/rest/services-reference/online/update-definition-feature-service-.htm}{a FeatureService}
\item deleting definitions for \href{https://developers.arcgis.com/rest/services-reference/online/delete-from-definition-feature-layer/}{a FeatureLayer} or a \href{https://developers.arcgis.com/rest/services-reference/online/delete-from-definition-feature-service/}{FeatureService}
}
}
\examples{
if (interactive()) {
# authenticate
set_arc_token(auth_code())

# publish a layer
published <- publish_layer(iris, "Iris Test")

iris_fl <- arc_open(published$services$encodedServiceURL) |>
  get_layer(0)

# Update the name of the layer
update_definition(
  iris_fl,
  name = "New Layer Name"
)

# add an index on the the layer
add_definition(
  penguin_fl,
  indexes = list(
    name = "index1",
    fields = "Species",
    isUnique = FALSE,
    isAscending = FALSE,
    description = "Example index"
  )
)

# refresh the layer to get the updates
iris_fl <- refresh_layer(pengiris_flin_fl)
iris_fl[["indexes"]]
}
}
