% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-features.R
\name{select.FeatureLayer}
\alias{select.FeatureLayer}
\alias{select.Table}
\alias{filter.FeatureLayer}
\alias{filter.Table}
\alias{collect.FeatureLayer}
\alias{dplyr}
\alias{collect.Table}
\title{dplyr methods}
\usage{
select.FeatureLayer(x, ...)

select.Table(x, ...)

filter.FeatureLayer(x, ...)

filter.Table(x, ...)

collect.FeatureLayer(x, n_max = Inf, token = Sys.getenv("ARCGIS_TOKEN"), ...)

collect.Table(x, n_max = Inf, token = Sys.getenv("ARCGIS_TOKEN"), ...)
}
\description{
dplyr methods
}
\details{
The Feature Layer method of \code{collect()} will overwrite the \code{returnGeometry} parameter if set with \code{update_params()}. Use the \code{geometry} argument in \code{collect()} to set the parameter.
}
\keyword{internal}
