% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-layer-class.R
\name{feature_layer}
\alias{feature_layer}
\title{Create a Feature Layer Object}
\usage{
feature_layer(url, token = Sys.getenv("ARCGIS_TOKEN"))
}
\arguments{
\item{url}{the url of a feature layer}

\item{token}{authentication toekn as provided by \code{auth_client()} or \code{auth_code()}}
}
\description{
Ideally, we can use this s3 object to build up queries
}
\examples{
library(arcgis)
furl <- "https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services/USA_Counties_Generalized_Boundaries/FeatureServer/0"
county_fl <- feature_layer(furl)
}
