% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-server.R
\name{get_layer}
\alias{get_layer}
\alias{get_all_layers}
\title{Extract a feature layer}
\usage{
get_layer(x, id = 0, token = Sys.getenv("ARCGIS_TOKEN"))

get_all_layers(x, token = Sys.getenv("ARCGIS_TOKEN"))
}
\value{
\code{get_layer()} returns a single \code{FeatureLayer} or \code{Table} based on its ID
\code{get_all_layers()} returns a named \code{list} with an element \code{layers} and \code{tables}. Each a list containing \code{FeatureLayer} and \code{Table}s respectively.
}
\description{
From a Feature Server, extract a feature layer.
}
