% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-server.R
\name{get_layer}
\alias{get_layer}
\alias{get_all_layers}
\title{Extract a feature layer}
\usage{
get_layer(x, id = 0, token = "")

get_all_layers(x, token = "")
}
\value{
`get_layer()` returns a single `FeatureLayer` or `Table` based on its ID
`get_all_layers()` returns a named `list` with an element `layers` and `tables`. Each a list containing `FeatureLayer` and `Table`s respectively.
}
\description{
From a Feature Server, extract a feature layer.
}
