% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-feature-server.R
\name{get_layer}
\alias{get_layer}
\alias{get_all_layers}
\title{Extract a layer from a FeatureServer}
\usage{
get_layer(x, id = 0, token = Sys.getenv("ARCGIS_TOKEN"))

get_all_layers(x, token = Sys.getenv("ARCGIS_TOKEN"))
}
\value{
\itemize{
\item \code{get_layer()} returns a single \code{FeatureLayer} or \code{Table} based on its ID
\item \code{get_all_layers()} returns a named \code{list} with an element \code{layers} and \code{tables}.
Each a list containing \code{FeatureLayer} and \code{Table}s respectively.
}
}
\description{
\code{FeatureServer}s are collections of \code{Table}s and \code{FeatureLayer}s. These helpers
provide easy access to the layers contained in the \code{FeatureServer}.
}
