% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_arc_content_url}
\alias{is_arc_content_url}
\title{Test if the vector is a content url of the supplied type}
\usage{
is_arc_content_url(
  x,
  type = c("search", "item", "group", "user"),
  ...,
  perl = TRUE
)
}
\description{
Test if the vector is a content url of the supplied type
}
\keyword{internal}
