% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster-fns.R
\name{list_service_raster_fns}
\alias{list_service_raster_fns}
\title{List Available Raster Funcitons}
\usage{
list_service_raster_fns(
  x,
  arg = rlang::caller_arg(x),
  call = rlang::caller_call()
)
}
\arguments{
\item{x}{an \code{ImageServer}.}
}
\value{
a data.frame of the available raster functions.
}
\description{
List Available Raster Funcitons
}
