% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc-select.R
\name{obj_check_layer}
\alias{obj_check_layer}
\title{Check if an object is a FeatureLayer or Table object}
\usage{
obj_check_layer(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{A \code{FeatureLayer} or \code{Table} class object created with \code{\link[=arc_open]{arc_open()}}.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
\code{\link[=obj_check_layer]{obj_check_layer()}} errors if an object does not inherit either the
FeatureLayer or Table class.
}
\keyword{internal}
