% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-imagery.R
\name{query_imagery}
\alias{query_imagery}
\title{Retrieve Imagery}
\usage{
query_imagery(
  x,
  bbox,
  format = "tiff",
  width = NULL,
  height = NULL,
  token = Sys.getenv("ARCGIS_TOKEN")
)
}
\description{
Retrieve Imagery
}
