% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spatial-filter.R
\name{prepare_spatial_filter}
\alias{prepare_spatial_filter}
\alias{match_spatial_rel}
\title{Prepare json for spatial filters}
\usage{
prepare_spatial_filter(filter_geom, crs, predicate)

match_spatial_rel(predicate)
}
\arguments{
\item{filter_geom}{an object of class \code{bbox}, \code{sfc} or \code{sfg} used to filter
query results based on a predicate function. If an \code{sfc} object is provided
it will be transformed to the layers spatial reference. If the \code{sfc} is
missing a CRS (or is an \code{sfg} object) it is assumed to be in the layers
spatial reference. If an \code{sfc} object has multiple features, the features
are unioned with \code{\link[sf:geos_combine]{sf::st_union()}}. If an \code{sfc} object has MULTIPOLYGON
geometry, the features are treated as polygonal coverage and unioned with
\code{is_coverage = TRUE} before being cast to POLYGON geometry with
\code{\link[sf:st_cast]{sf::st_cast()}}.}

\item{predicate}{default \code{"intersects"}. Possible options are \code{"intersects"},  \code{"contains"},  \code{"crosses"},  \code{"overlaps"},  \code{"touches"}, and \code{"within"}.
\subsection{Spatial Binary Predicates:}{
\itemize{
\item esriSpatialRelIntersects
\item esriSpatialRelContains
\item esriSpatialRelCrosses
\item esriSpatialRelOverlaps
\item esriSpatialRelTouches
\item esriSpatialRelWithin
}
}}
}
\description{
Prepare json for spatial filters
}
