% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-attachments.R
\name{update_attachments}
\alias{update_attachments}
\title{Update Feature Service Attachments}
\usage{
update_attachments(
  x,
  feature_id,
  attachment_id,
  path,
  .progress = TRUE,
  token = arc_token()
)
}
\arguments{
\item{feature_id}{a vector of object IDs that corresponds to the feature of the corresponding \code{attachment_id}.}

\item{attachment_id}{the ID of the attachment—this corresponds to the \code{id} column returned from \code{query_layer_attachments()}}

\item{token}{an \code{httr2_token} as created by \code{auth_code()} or similar}
}
\value{
a \code{data.frame} with 2 columns returning the status of the update.
}
\description{
Feature Services can contain attachments that are associated with a single feature ID. This function enables you to update the attachments of multiple features at once by generating multiple requests and performing them in parallel.
}
\details{
To rename or otherwise an attachment in a Feature Service, you must first download that attachment, modify the file on disk, and then upload it again. This is a limitation of ArcGIS Online and Enterprise. If you'd like to see this changed, please submit a community idea at \href{https://community.esri.com/t5/arcgis-online/ct-p/arcgis-online}{community.esri.com}.
}
\references{
See \href{https://developers.arcgis.com/rest/services-reference/enterprise/update-attachment/#request-parameters}{API documentation} for more.
}
