% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_definition.R
\name{update_definition}
\alias{update_definition}
\alias{.update_definition_name}
\alias{.update_layer_properties}
\title{Update Feature Layer or Service definition}
\usage{
update_definition(x, update = NULL, async = FALSE, token = arc_token())

.update_definition_name(...)

.update_layer_properties(
  description = NULL,
  license_text = NULL,
  xss_trusted_fields = NULL,
  display_field = NULL,
  min_scale = NULL,
  max_scale = NULL,
  max_records = NULL,
  default_visibility = NULL,
  ...
)
}
\arguments{
\item{x}{A Feature Layer or Feature Service class object.}

\item{update}{A named list used as the "updateDefinition" body for the
request.}

\item{async}{Default \code{FALSE}. If \code{TRUE}, support asynchronous processing for
the request.}
}
\description{
\url{https://developers.arcgis.com/rest/services-reference/online/update-definition-feature-layer-.htm#ESRI_SECTION2_027B8A04953C409E9A9D9051DD4CDBEE}
\url{https://developers.arcgis.com/rest/services-reference/online/update-definition-feature-service-.htm}
}
