% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect-layer.R
\name{update_params}
\alias{update_params}
\title{Modify Query Parameters}
\usage{
update_params(x, ...)
}
\arguments{
\item{x}{a `FeatureLayer` object}

\item{...}{key value pairs of query parameters and values.}
}
\description{
Modify Query Parameters
}
