% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clear_query}
\alias{clear_query}
\alias{list_fields}
\alias{list_items}
\alias{refresh_layer}
\title{Utility functions}
\usage{
clear_query(x)

list_fields(x)

list_items(x)

refresh_layer(x)
}
\arguments{
\item{x}{A \code{FeatureLayer} or \code{Table} class object created with \code{\link[=arc_open]{arc_open()}}.}
}
\value{
See Details.
}
\description{
Utility functions
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\itemize{
\item \code{list_fields()} returns a data.frame of the fields in a \code{FeatureLayer} or \code{Table}
\item \code{list_items()} returns a data.frame containing the layers or tables in a \code{FeatureServer} or \code{MapServer}
\item \code{clear_query()} removes any saved query in a \code{FeatureLayer} or \code{Table} object
\item \code{refresh_layer()} syncs a \code{FeatureLayer} or \code{Table} with the remote
resource picking up any changes that may have been made upstream.
Returns an object of class \code{x}.
}
}
